introduce enum class SensorType
better type safety, avoid flipping sensor index with sensor type
in device.cpp:
- use static_cast<uint8_t> in dsm-api conversions, use static_cast<int>
  when converting to pass it via event structure internally inside
  the dss
- replace intToString by sensorName in log messages to avoid casting,
  and where I assume the log is not processed automatically

+++ src/web/handler/jsonhelper.cpp
@@ -172,7 +172,7 @@ namespace dss {
     const std::vector<boost::shared_ptr<DeviceSensor_t> > sensors = _device.getDevice()->getSensors();
     for (size_t i = 0; i < sensors.size(); i++) {
       _json.startObject();
-      _json.add("type", sensors.at(i)->m_sensorType);
+      _json.add("type", static_cast<int>(sensors.at(i)->m_sensorType));
       _json.add("valid", sensors.at(i)->m_sensorValueValidity);
       _json.add("value", sensors.at(i)->m_sensorValueFloat);
       _json.endObject();
